+++ src/App.js
@@ -5,7 +5,7 @@ import './App.css';
 import Navbar from './components/Navbar/Navbar';
 import { Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
-import SinglePosts from './pages/SinglePosts';
+import SinglePosts from './pages/SinglePost';
 import BlogPosts from './pages/BlogPosts';
 import Projects from './pages/Projects';
 import Error from './pages/Error';
