+++ src/components/SinglePostContent/SinglePostContent.module.css
@@ -1,0 +1,13 @@
+.container_main_image{
+  width: 100vw;
+  height: 250px;
+  max-width: 100%;
+  overflow: hidden;
+}
+
+.post_main_image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  object-position: center;
+}
